+++ src/Sentry/Laravel/Tracing/EventHandler.php
@@ -11,7 +11,6 @@ use Illuminate\Queue\Events as QueueEvents;
 use Illuminate\Queue\Queue;
 use Illuminate\Queue\QueueManager;
 use Illuminate\Routing\Events as RoutingEvents;
-use Illuminate\Routing\Route;
 use RuntimeException;
 use Sentry\Laravel\Integration;
 use Sentry\SentrySdk;
@@ -32,11 +31,8 @@ class EventHandler
      * @var array
      */
     protected static $eventHandlerMap = [
-        'router.matched' => 'routerMatched',                    // Until Laravel 5.1
-        RoutingEvents\RouteMatched::class => 'routeMatched',    // Since Laravel 5.2
-
-        'illuminate.query' => 'query',                          // Until Laravel 5.1
-        DatabaseEvents\QueryExecuted::class => 'queryExecuted', // Since Laravel 5.2
+        RoutingEvents\RouteMatched::class => 'routeMatched',
+        DatabaseEvents\QueryExecuted::class => 'queryExecuted',
     ];
 
     /**
@@ -45,9 +41,9 @@ class EventHandler
      * @var array
      */
     protected static $queueEventHandlerMap = [
-        QueueEvents\JobProcessing::class => 'queueJobProcessing',               // Since Laravel 5.2
-        QueueEvents\JobProcessed::class => 'queueJobProcessed',                 // Since Laravel 5.2
-        QueueEvents\JobExceptionOccurred::class => 'queueJobExceptionOccurred', // Since Laravel 5.2
+        QueueEvents\JobProcessing::class => 'queueJobProcessing',
+        QueueEvents\JobProcessed::class => 'queueJobProcessed',
+        QueueEvents\JobExceptionOccurred::class => 'queueJobExceptionOccurred',
     ];
 
     /**
@@ -129,8 +125,6 @@ class EventHandler
      * Attach all event handlers.
      *
      * @uses self::routeMatchedHandler()
-     * @uses self::routerMatchedHandler()
-     * @uses self::queryHandler()
      * @uses self::queryExecutedHandler()
      */
     public function subscribe(): void
@@ -152,8 +146,8 @@ class EventHandler
      *
      * @param \Illuminate\Queue\QueueManager $queue
      *
-     * @uses self::queueJobProcessedHandler()
      * @uses self::queueJobProcessingHandler()
+     * @uses self::queueJobProcessedHandler()
      * @uses self::queueJobExceptionOccurredHandler()
      */
     public function subscribeQueueEvents(QueueManager $queue): void
@@ -163,19 +157,16 @@ class EventHandler
             return;
         }
 
-        // The payload create callback was introduced in Laravel 5.7 so we need to guard against older versions
-        if (method_exists(Queue::class, 'createPayloadUsing')) {
-            Queue::createPayloadUsing(static function (?string $connection, ?string $queue, ?array $payload): ?array {
-                $currentSpan = Integration::currentTracingSpan();
+        Queue::createPayloadUsing(static function (?string $connection, ?string $queue, ?array $payload): ?array {
+            $currentSpan = Integration::currentTracingSpan();
 
-                if ($currentSpan !== null && $payload !== null) {
-                    $payload[self::QUEUE_PAYLOAD_TRACE_PARENT_DATA] = $currentSpan->toTraceparent();
-                    $payload[self::QUEUE_PAYLOAD_BAGGAGE_DATA] = $currentSpan->toBaggage();
-                }
+            if ($currentSpan !== null && $payload !== null) {
+                $payload[self::QUEUE_PAYLOAD_TRACE_PARENT_DATA] = $currentSpan->toTraceparent();
+                $payload[self::QUEUE_PAYLOAD_BAGGAGE_DATA] = $currentSpan->toBaggage();
+            }
 
-                return $payload;
-            });
-        }
+            return $payload;
+        });
 
         $queue->looping(function () {
             $this->afterQueuedJob();
@@ -214,12 +205,7 @@ class EventHandler
         }
     }
 
-    /**
-     * Until Laravel 5.1
-     *
-     * @param Route $route
-     */
-    protected function routerMatchedHandler(Route $route): void
+    protected function routeMatchedHandler(RoutingEvents\RouteMatched $match): void
     {
         $transaction = Integration::currentTransaction();
 
@@ -227,52 +213,13 @@ class EventHandler
             return;
         }
 
-        [$transactionName, $transactionSource] = Integration::extractNameAndSourceForRoute($route);
+        [$transactionName, $transactionSource] = Integration::extractNameAndSourceForRoute($match->route);
 
         $transaction->setName($transactionName);
         $transaction->getMetadata()->setSource($transactionSource);
     }
 
-    /**
-     * Since Laravel 5.2
-     *
-     * @param \Illuminate\Routing\Events\RouteMatched $match
-     */
-    protected function routeMatchedHandler(RoutingEvents\RouteMatched $match): void
-    {
-        $this->routerMatchedHandler($match->route);
-    }
-
-    /**
-     * Until Laravel 5.1
-     *
-     * @param string $query
-     * @param array  $bindings
-     * @param int    $time
-     * @param string $connectionName
-     */
-    protected function queryHandler($query, $bindings, $time, $connectionName): void
-    {
-        $this->recordQuerySpan($query, $time);
-    }
-
-    /**
-     * Since Laravel 5.2
-     *
-     * @param \Illuminate\Database\Events\QueryExecuted $query
-     */
     protected function queryExecutedHandler(DatabaseEvents\QueryExecuted $query): void
-    {
-        $this->recordQuerySpan($query->sql, $query->time);
-    }
-
-    /**
-     * Helper to add an query breadcrumb.
-     *
-     * @param string     $query
-     * @param float|null $time
-     */
-    private function recordQuerySpan($query, $time): void
     {
         if (!$this->traceSqlQueries) {
             return;
@@ -287,9 +234,9 @@ class EventHandler
 
         $context = new SpanContext();
         $context->setOp('db.sql.query');
-        $context->setDescription($query);
-        $context->setStartTimestamp(microtime(true) - $time / 1000);
-        $context->setEndTimestamp($context->getStartTimestamp() + $time / 1000);
+        $context->setDescription($query->sql);
+        $context->setStartTimestamp(microtime(true) - $query->time / 1000);
+        $context->setEndTimestamp($context->getStartTimestamp() + $query->time / 1000);
 
         if ($this->traceSqlQueryOrigins) {
             $queryOrigin = $this->resolveQueryOriginFromBacktrace();
@@ -320,11 +267,6 @@ class EventHandler
         return "{$filePath}:{$firstAppFrame->getLine()}";
     }
 
-    /**
-     * Since Laravel 5.2
-     *
-     * @param \Illuminate\Queue\Events\JobProcessing $event
-     */
     protected function queueJobProcessingHandler(QueueEvents\JobProcessing $event): void
     {
         $parentSpan = Integration::currentTracingSpan();
@@ -390,21 +332,11 @@ class EventHandler
         SentrySdk::getCurrentHub()->setSpan($this->currentQueueJobSpan);
     }
 
-    /**
-     * Since Laravel 5.2
-     *
-     * @param \Illuminate\Queue\Events\JobExceptionOccurred $event
-     */
     protected function queueJobExceptionOccurredHandler(QueueEvents\JobExceptionOccurred $event): void
     {
         $this->afterQueuedJob(SpanStatus::internalError());
     }
 
-    /**
-     * Since Laravel 5.2
-     *
-     * @param \Illuminate\Queue\Events\JobProcessed $event
-     */
     protected function queueJobProcessedHandler(QueueEvents\JobProcessed $event): void
     {
         $this->afterQueuedJob(SpanStatus::ok());