+++ test/Sentry/Laravel/LogChannelTest.php
@@ -2,7 +2,6 @@
 
 namespace Sentry\Laravel\Tests\Sentry\Laravel;
 
-use Illuminate\Log\LogManager;
 use Monolog\Handler\FingersCrossedHandler;
 use Sentry\Laravel\LogChannel;
 use Sentry\Laravel\SentryHandler;
@@ -12,8 +11,6 @@ class LogChannelTest extends SentryLaravelTestCase
 {
     public function test_creating_handler_without_action_level_config()
     {
-        $this->skipIfLogManagerNotAvailable();
-
         $logChannel = new LogChannel($this->app);
         $logger = $logChannel([]);
 
@@ -22,8 +19,6 @@ class LogChannelTest extends SentryLaravelTestCase
 
     public function test_creating_handler_with_action_level_config()
     {
-        $this->skipIfLogManagerNotAvailable();
-
         $logChannel = new LogChannel($this->app);
         $logger = $logChannel(['action_level' => 'critical']);
 
@@ -35,14 +30,5 @@ class LogChannelTest extends SentryLaravelTestCase
         $loggerWithoutActionLevel = $logChannel(['action_level' => null]);
 
         $this->assertContainsOnlyInstancesOf(SentryHandler::class, $loggerWithoutActionLevel->getHandlers());
-    }
-
-    private function skipIfLogManagerNotAvailable()
-    {
-        if (class_exists(LogManager::class)) {
-            return;
-        }
-
-        $this->markTestSkipped('Laravel version <=5.5 does not contain the LogManager required for this functionality.');
     }
 }